Adapt the Transaction::verify API to not require mutations
I think it makes more sense for the verify method to keep track of
double-spending itself instead of requiring the closure to do that.
When internally keeping track of double-spending, the closure can just
be a simple inspector closure and doesn't need to mutate any state,
making it a lot simpler and potentially more efficient.

Also, OutPoint is Copy, so doesn't need to be referenced.