Make secp256k1 optional
This change makes `secp256k1` an optional dependency so that
applications requiring chain parsing but not doing any validation nor
signing do not have to compile `secp256k1`. This is a breaking change
because dependencies of this crate must fix their feature settings but
apart from that they don't need to do anything.

There are two things which may be surprising: moving of some modules and
removing `secp256k1/serde` from `use-serde`. The former one is explained
in the doc of `src/util/psbt_no_key/mod.rs`. The latter is due to
inability to avoid compiling `secp256k1` when not needed. (`electrs` hit
this)