Allow unknown lints
We support many versions of rustc by design, this puts us in the
following predicament:

1. Linting with a newer version of clippy throws a warning for a newly
introduced default lint.
2. In order to allow the lint we must use the lint by name but this lint
does not exist in earlier versions of clippy, so we get a warning
`unknown lint` when linting with the earlier version.

We want to have our cake and eat it too, we can do this by allowing
unknown lints. Doing so enables (1) while preventing (2).

This will help us introduce clippy into the CI pipeline because we can
better control lint output for different versions of the toolchain.

Before this patch was applied `cargo +1.36 check` emits:

    warning: unknown lint: `broken_intra_doc_links`

With this patch applied the warning is removed.