Add plural and singular for bits/sats
Currently we use plural for denomination "bits" and singular for
denomination "sat". This means our users must remember which to use
which is annoying.

There is no reason not to allow parsing plural and singular, at least
for the satoshi abbreviation "sats" is a common colloquial term.

Enable usage of "bits", "bit", "sat", and "sats" when parsing and
displaying denomination.