Enable deriving serde using as_str module
We recently added a `parse` function to enable deserialization of
strings for `Amount` within a struct. For sats and BTC we have a module
for this so we can go both ways, unless I'm missing something there
doesn't seem to be a reason to not allow serialization of `Amount`
strings to include the denomination also.

With this applied one can do the following:
```
        #[derive(Deserialize, Serialize, PartialEq, Debug)]
        struct T {
            #[serde(default, with = "::util::amount::serde::as_string")]
            pub amt: Amount,
        }
```