Add a workspace
Add a workspace to the top level directory.

Create a directory `bitcoin` and move into it the following as is with
no code changes:

- src
- Cargo.toml
- test_data
- examples

Add a workspace to the repository root directory. Add the newly created
`bitcoin` crate to the workspace.

Fix `contrib/test.sh`, fuzz, embedded, and CI so everything still passes.