Add ChainHash type
The Lightning network defines a type called 'chain hash' that is used to
uniquely represent the various Bitcoin networks as a 32 byte hash value.
It is calculated by hashing the genesis block for the respective
network. Chain hash is now being used by the DLC folks, as such it is
useful to have it implemented in `rust-bitcoin`.

Add a `ChainHash` type that can be used to get the unique identifier of
each of the 4 Bitcoin networks we support.