Add ScriptRef<'a> type
`Script` is a read-only data type but currently we have no way of
creating one from a slice without copying the inner byte array.

It will be un-ergonomic if we use a reference inside of `Script` so
instead add a `ScriptRef<'a>` type that can be used when we know we only
need a reference and never need to take ownership of the inner array.