Use assert_eq for len check
According to the Rust docs for `debug_assert` it should only be used if
there a proved performance costs to using `assert`.

Checking the length of a `BinaryHeap` is equivalent to checking the
length of a `Vec` which is fast, therefore its unlikely that using
`assert_eq` will have any noticeable performance costs.