Add cfg nightly
Currently we use

 `#![cfg_attr(all(test, feature = "unstable"), feature(test))]`

in `lib.rs` to enable unstable test features. This works but has the
unfortunate side effect that we can not use `cargo test --all-features`
because this command enables `"unstable"` and `test` but uses the stable
toolchain. The problem is further compounded by the fact that we cannot
use `cargo clippy --all-features` so we cannot lint the test code.

There is a `nightly` feature that can be used with `cfg` to get around
this, since enabling the `"unstable"` feature implies that we are using
a nightly toolchain this has no ill-effects.

Add `nightly` to the compiler attribute when guarding on the
`"unstable"` feature. E.g.,

 `#[cfg(all(nightly, test, feature = "unstable"))]`