Add a test to check bitcoin_hashes dependency
Recently we updated the `bitcoin_hashes` dependency without first doing
so in `secp256k1`. Since `bitcoin_hashes` is an optional dependency of
secp this did not show up in our CI run and was reported by a user.

To prevent this from happening again we can add a test that creates two
arbitrary objects, one from the `rust-bitcoin` `bitcoin_hashes`
dependency and one from the `rust-secp256k1` `bitcoin_hashes` dependency
then assert that they are comparable. If different versions of
`bitcoin_hashes` are being used then the compiler will complain.