Add ensure_no_std crate
Currently we make a bunch of attempts at enabling users in a `no_std`
environment to use `rust-bitcoin`, so far we are failing at this
endevour. In an attempt to get `rust-bitcoin` building in a `no_std`
environment add a crate `ensure_no_std` that specifically checks if a
dependency can be build in a `no_std` environment.

Adds zero dependencies, so currently proves nothing. We will want to
eventually add all our dependencies to this crate (incl. `rust-bitcoin`).