Introduce new canonical license blurb
Recently we introduced SPDX license ids but there was some confusion
about the use, form, and purpose of lines before (library name, author).

It was pointed out to me by Andrew that if/when folks cut'n'paste our
code they often will keep the whole blurb. Design it with that in
mind. FTR, Andrew also indicated that he did not mind his name being
removed in favour of "The rust-bitcoin developers".

At first I thought the term "The rust-bitcoin developers" was a bit
wishy-washy but yesterday I realised that I am proud to be part of that
crew, striving to deliver code to the highest possible standard.

Introduce a canonical format for the licences blurb.

- Use the name of the library
- Use "The rust-bitcoin developers"
- Use the SPDX ID