Optionally re-export the bech32 crate
Add a re-export statement for the `bech32` crate, the aim is to keep all
the bitcoin stringly encodings in one place for users. Make the
dependency optional to keep the dependency graph lean. In order to do
this and enable folks to correctly use "std" feature add a "bech32-std"
feature and comment the manifest liberally.

Note, the rustdocs do not show up in HTML rendered docs, would be nice
if they did.