Optionally re-export base64 crate
Add a re-export statement for the `base64` crate, the aim is to keep all
the bitcoin stringly encodings in one place for users. Make the
dependency optional to keep the dependency graph lean. In order to do
this and enable folks to correctly use "std" feature add a "base64-std"
feature and comment the manifest liberally.