Use sighash::SighashCache for bip143 unit tests
We have a bunch of unit tests that verify the BIP 143 test vectors.
Three of these use the deprecated `SighashComponents` struct. We should
implement these tests using the new `SighashCache`. In order to do so we
have to move the tests to the `sighash` module so they can get access to
the private `segwit_cache()` function, and verify the cache internal
state against the test vectors.

Re-write the BIP 143 test vector code using `SighashCache` and remove
the versions using `SighashComponents`.

Done in preparation for removing the deprecated `SighashComponents`.