Add newtype TxAmount
This type is used as the value field of `TxOut` and has an invariant
that its value is never greater than 21 million.

There is one complexity: the u64::MAX (all bits on) value is apparently
used to signal in consensus code "null txout". This is currently handled
by having `TxAmount` be an enum - its a bit ugly though.