Deprecate SignedAmount
We currently provide a very basic signed integer type `SignedAmount`. It
is not obvious what this type is for. On top of that it only provides a
primitive subset of integer operations.

Also, there is ongoing discussion about what `Amount` is even for. Is it
a type to represent the value that a single utxo can have? Is it a type
that can represent a total transaction amount? Is it a type for summing
all bitcoin moved by all transactions for all time? Is it for parsing
and pretty printing bitcoin amounts (sats and BTC)? Its not clear.

It seems that the `SignedAmount` is either unneeded or incomplete. Lets
deprecate it (and later remove it) to save us the maintenance burden.