Add integer serialization tests
Add some negative integer unit tests and remove the TODO.

Note, this is not really that necessary because we are going to move to
using the stdlib `to_le_bytes` methods so we don't really need any of
these test except to test that we are calling the correct be/le method.
They do no harm however.