Move base58 to its own crate
We want to create a `bitcoin-str` crate where all the encodings live,
however base58 presents a problem because it depends on `hashes` (and
`hex`) but `hex` will eventually live in `bitcoin-str` so this
introduces a circular dependency.

The solution is that either `base58` has to be its own crate or `hex`
has to be, elect to make `base58` its own crate because it is less
important to the rest of the ecosystem (i.e., its not used that much
compared to `hex`).