Implement conversion traits for lock types
When we originally added `absolute::LockTime` and `Sequence` we decided
to be conservative and only provide constructor methods with more
descriptive names.

`Sequence` has a public inner field so there is not a lot of clarity
gained between `Sequence(x)` and `Sequence::from(x)`.

As for `absolute::LockTime`, a `From` impl can call `from_consensus`
providing a more ergonomic conversion method at the risk of being
marginally less clear. We aim to make it hard for users of the library
to misuse the library but we should not do so at the cost of ergonomics
unless there is a _really_ clear benefit, in this instance that is not
the case.

Add conversion methods `From` for the two types mentioned above. Also
add `TryFrom` impls for `Height` and `Time` for similar reasons.